TECH Split createWorker and createWorkers functions
