+++ lib/createWorkers.js
@@ -5,55 +5,18 @@ const co = require('co');
 const url = require('url');
 const amqplib = require('amqplib');
 const logger = require('chpr-logger');
+
+const {
+  promisifyWithTimeout,
+  subscribeToConnectionEvents,
+  subscribeToChannelEvents
+} = require('./lib');
 const {
   applyConfiguration,
-  applyOptions,
-  applyLegacyHandler,
-  applyLegacyConfig,
-  applyLegacyOptions
-} = require('./schema');
+  applyOptions
+} = require('./schema/createWorkers.schema');
 
 const DEFAULT_EXCHANGE_TYPE = 'topic';
-
-function formatLegacyParameters(handler, config, options) {
-  return {
-    handlers: [{ handle: handler, validate: options.validator, routingKey: config.routingKey }],
-    config: _.omit(config, 'routingKey'),
-    options: _.omit(options, 'validator')
-  };
-}
-
-/**
- * Create a worker instance based on configuration
- * @param {function} handler function to handle each message
- * @param {Object} config configuration for the worker
- * @param {String} config.workerName name for the worker
- * @param {String} config.amqpUrl url to amqp server
- * @param {String} config.exchangeName name of the exchange to use
- * @param {String} config.queueName name of the queue to listen on
- * @param {String} [config.routingKey] name of the routing key to bind to
- * @param {Object} [options] additional options
- * @param {function} [options.validator] function to validate the message with
- * @param {Number} [options.heartbeat] to override default heartbeat
- * @param {Number} [options.taskTimeout] to override default task timeout
- * @param {Number} [options.processExitTimeout] to override default process exit timeout
- * @param {Number} [options.channelPrefetch] to override default channel prefetch value
- * @returns {Object} a worker instance with connection, channel, and listen/close functions
- * @deprecated
- */
-function createWorker(handler, config, options = {}) {
-  handler = applyLegacyHandler(handler);
-  config = applyLegacyConfig(config);
-  options = applyLegacyOptions(options);
-
-  const {
-    handlers,
-    config: newConfig,
-    options: newOptions
-  } = formatLegacyParameters(handler, config, options);
-
-  return createWorkers(handlers, newConfig, newOptions);
-}
 
 /**
  * Create a worker instance based on configuration
@@ -208,7 +171,7 @@ function createWorkers(handlers, config, options = {}) {
   function* _handleMessage(handler, content, fields) {
     const { workerName, taskTimeout } = configuration;
     try {
-      yield _promisifyWithTimeout(handler(content, fields), workerName, taskTimeout);
+      yield promisifyWithTimeout(handler(content, fields), workerName, taskTimeout);
       return true;
     } catch (err) {
       if (!fields.redelivered) {
@@ -243,7 +206,7 @@ function createWorkers(handlers, config, options = {}) {
       const urlWithParam = url.resolve(amqpUrl, `?heartbeat=${heartbeat}`);
       const connection = yield amqplib.connect(urlWithParam, socketOptions);
       logger.info({ workerName, heartbeat }, '[AMQP] connected to server');
-      _subscribeToConnectionEvents(connection, workerName);
+      subscribeToConnectionEvents(connection, workerName);
       return connection;
     } catch (err) {
       logger.error(
@@ -272,7 +235,7 @@ function createWorkers(handlers, config, options = {}) {
       { exchangeName, queueName, routingKey, channelPrefetch, workerName },
       '[AMQP] create channel'
     );
-    _subscribeToChannelEvents(
+    subscribeToChannelEvents(
       channel,
       { exchangeName, queueName, routingKey, channelPrefetch, workerName }
     );
@@ -307,72 +270,6 @@ function createWorkers(handlers, config, options = {}) {
   }
 }
 
-/**
- * Make a promise outside of a yieldable.
- * To avoid unhandled callbacks resulting in unacked messages piling up in amqp,
- * the returned promise will fail after a specified timeout.
- * @param {Promise|Generator|Function} yieldable a yieldable (generator or promise)
- * @param {String} name  Name for error handling
- * @param {Number} timeout Time after which the promise fails
- * @returns {Promise} the promise
- */
-function _promisifyWithTimeout(yieldable, name, timeout) {
-  return new Promise((resolve, reject) => {
-    co(yieldable).then(resolve).catch(reject);
-    setTimeout(() => reject(new Error(`Yieldable timeout in ${name}`)), timeout);
-  });
-}
-
-/**
- * Bind a logger to noticeable connection events
- * @param {Object} connection the connection objext
- * @param {String} workerName worker name
- * @returns {*} -
- * @private
- */
-function _subscribeToConnectionEvents(connection, workerName) {
-  connection.on('close', forceExit => {
-    logger.info({ workerName, forceExit }, '[AMQP] Connection closing, exiting');
-  });
-  connection.on('error', err => {
-    logger.error({ err, workerName }, '[AMQP] Connection closing because of an error');
-  });
-  connection.on('blocked', reason => {
-    logger.warn({ reason, workerName }, '[AMQP] Connection blocked');
-  });
-}
-
-/**
- * Bind a logger to noticeable channel events
- * @param {Object} channel the channel object
- * @param {String} exchangeName the exchange name to use
- * @param {String} queueName the queue name to target
- * @param {String} routingKey routing key to bind on
- * @param {Number} channelPrefetch number of message to prefetch from channel
- * @param {String} workerName the name of the worker (for logs)
- * @returns {*} -
- * @private
- */
-function _subscribeToChannelEvents(channel,
-  { exchangeName, queueName, routingKey, channelPrefetch, workerName }) {
-  channel.on('error', err => {
-    logger.error(
-      { err, exchangeName, queueName, routingKey, channelPrefetch, workerName },
-      '[AMQP] channel error'
-    );
-  });
-  channel.on('close', () => {
-    logger.info(
-      { exchangeName, queueName, routingKey, channelPrefetch, workerName },
-      '[AMQP] channel closed'
-    );
-  });
-}
-
 module.exports = {
-  createWorker,
-  createWorkers,
-  _subscribeToConnectionEvents,
-  _subscribeToChannelEvents,
-  _promisifyWithTimeout
+  createWorkers
 };
